完善 aof 和 bg aof

+++ syncio.go
@@ -35,7 +35,7 @@ func fwriteBulkObject(fp *os.File, o *Gobj) int8 {
 	}
 	return GODIS_ERR
 }
-func fwriteBulkCount(fp *os.File, char byte, count int64) int8 {
+func fwriteBulkCount(fp *os.File, char byte, count int) int8 {
 	s := fmt.Sprintf("%c%v\r\n", char, count)
 	if _, err := fp.WriteString(s); err != nil {
 		log.Printf("Error writing bulk count to file: %v\n", err)
